pyproject.toml
[project]
name = "payout-service"
version = "0.1.0"
description = ""
authors = [
    {name = "loe",email = "Law_Of_Evi1@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==4.2.*)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-django = "^4.11.1"
black = "^26.1.0"
isort = "^7.0.0"
flake8 = "^7.3.0"

